src/robot_control/control_example.py
#!/usr/bin/env python3

import argparse
import logging
import time
from typing import Any, Dict, List, Optional

import pybullet as p
import pybullet_data
import yaml


class RobotArm:
    """机械臂控制类"""

    def __init__(self, config_path: str):
        """
        初始化机械臂控制器

        Args:
            config_path: 配置文件路径
        """
        self.config = self._load_config(config_path)
        self._setup_logging()
        self._setup_physics()
        self._load_robot()

    def _load_config(self, config_path: str) -> Dict[str, Any]:
        """加载配置文件"""
        with open(config_path) as f:
            return yaml.safe_load(f)["robot_control"]

    def _setup_logging(self):
        """设置日志系统"""
        log_config = self.config["logging"]
        logging.basicConfig(
            level=getattr(logging, log_config["level"]),
            format="%(asctime)s - %(name)s - %(levelname)s - %(message)s",
            handlers=[
                logging.FileHandler(log_config["file"]),
                logging.StreamHandler() if log_config["console_output"] else logging.NullHandler(),
            ],
        )
        self.logger = logging.getLogger(__name__)

    def _setup_physics(self):
        """设置物理引擎"""
        # 连接到物理引擎
        self.client = p.connect(p.GUI if self.config["visualization"]["enabled"] else p.DIRECT)
        p.setAdditionalSearchPath(pybullet_data.getDataPath())
        p.setGravity(0, 0, -9.81)

        # 设置物理引擎参数
        p.setTimeStep(self.config["physics"]["time_step"])
        p.setRealTimeSimulation(False)

    def _load_robot(self):
        """加载机器人模型"""
        robot_config = self.config["robot"]
        self.robot_id = p.loadURDF(
            robot_config["urdf_path"],
            basePosition=robot_config["base_position"],
            baseOrientation=p.getQuaternionFromEuler(robot_config["base_orientation"]),
            useFixedBase=True,
        )

        # 获取关节信息
        self.num_joints = p.getNumJoints(self.robot_id)
        self.joint_info = {}
        for i in range(self.num_joints):
            info = p.getJointInfo(self.robot_id, i)
            self.joint_info[info[1].decode("utf-8")] = {
                "id": info[0],
                "type": info[2],
                "lower_limit": info[8],
                "upper_limit": info[9],
                "max_force": info[10],
                "max_velocity": info[11],
            }

        self.logger.info(f"Robot loaded with {self.num_joints} joints")

    def get_joint_positions(self) -> List[float]:
        """获取当前关节位置"""
        positions = []
        for i in range(self.num_joints):
            joint_state = p.getJointState(self.robot_id, i)
            positions.append(joint_state[0])
        return positions

    def set_joint_positions(self, positions: List[float], max_force: Optional[float] = None):
        """
        设置关节位置

        Args:
            positions: 目标关节位置列表
            max_force: 最大关节力矩，如果为None则使用默认值
        """
        for i, pos in enumerate(positions):
            if max_force is None:
                max_force = self.joint_info[list(self.joint_info.keys())[i]]["max_force"]
            p.setJointMotorControl2(
                self.robot_id, i, p.POSITION_CONTROL, targetPosition=pos, force=max_force
            )

    def move_to_pose(self, target_pose: List[float], max_steps: int = 1000):
        """
        移动到目标位姿

        Args:
            target_pose: 目标位姿 [x, y, z, roll, pitch, yaw]
            max_steps: 最大步数
        """
        # 这里需要实现逆运动学求解
        # 示例中仅作简单演示
        current_pos = self.get_joint_positions()
        target_joints = self._inverse_kinematics(target_pose)

        if target_joints is None:
            self.logger.error("Failed to solve inverse kinematics")
            return False

        # 执行运动
        for step in range(max_steps):
            self.set_joint_positions(target_joints)
            p.stepSimulation()

            # 检查是否到达目标位置
            current_pos = self.get_joint_positions()
            if self._check_position_reached(current_pos, target_joints):
                self.logger.info("Target position reached")
                return True

            time.sleep(self.config["physics"]["time_step"])

        self.logger.warning("Failed to reach target position within max steps")
        return False

    def _inverse_kinematics(self, target_pose: List[float]) -> Optional[List[float]]:
        """
        逆运动学求解

        Args:
            target_pose: 目标位姿 [x, y, z, roll, pitch, yaw]

        Returns:
            关节角度列表，如果求解失败则返回None
        """
        # 这里需要实现具体的逆运动学算法
        # 示例中返回一个简单的解
        return [0.0] * self.num_joints

    def _check_position_reached(
        self, current: List[float], target: List[float], tolerance: float = 0.01
    ) -> bool:
        """检查是否到达目标位置"""
        return all(abs(c - t) < tolerance for c, t in zip(current, target))

    def cleanup(self):
        """清理资源"""
        p.disconnect(self.client)
        self.logger.info("Robot control system cleaned up")


def main():
    parser = argparse.ArgumentParser(description="Robot Arm Control")
    parser.add_argument("--config", type=str, required=True, help="Path to configuration file")
    args = parser.parse_args()

    robot = RobotArm(args.config)
    try:
        # 示例：移动到目标位置
        target_pose = [0.5, 0.0, 0.5, 0.0, 0.0, 0.0]  # [x, y, z, roll, pitch, yaw]
        robot.move_to_pose(target_pose)
    finally:
        robot.cleanup()


if __name__ == "__main__":
    main()
